renamed files to shorter names now that their application is clear from the package context
