+++ src/svg/elements.py
@@ -6,8 +6,8 @@ from __future__ import annotations
 
 import xml.etree.ElementTree as ET
 
-import svg.svg_validators as sv
-import svg.svg_parsers as sp
+import svg.validators as sv
+import svg.parsers as sp
 import trigonometry as trig
