+++ src/svg/enum_color_keywords.py
