+++ src/svg/parsers.py
@@ -3,8 +3,9 @@ other formats
 """
 
 import re
+
 import trigonometry as trig
-import svg.svg_validators as sv
+import svg.validators as sv
 
 def parse_coordinate_string(coordinate: str) -> list:
     """Uses re to figure out what string of coordinates are in path
