+++ tests/test_freecad_svg_file.py
@@ -7,16 +7,15 @@ import xml.etree.ElementTree as ET
 
 sys.path.append('src')
 
-import translators.freecad_svg_file as fcsf
 from translators.freecad_svg_file import SketchSVG
 
 import file_handlers
-from freecad.free_cad_object_wrappers import File as FreeCADFile
-from freecad.free_cad_object_wrappers import Sketch
-import freecad.freecad_sketch_readers as fsr
+from freecad.object_wrappers import File as FreeCADFile
+from freecad.object_wrappers import Sketch
+import freecad.sketch_readers as fsr
 
-from svg.svg_file import SVGFile
-import svg.svg_generators as sg
+from svg.file import SVGFile
+import svg.generators as sg
 
 class TestFreeCADSVGFile(unittest.TestCase):
